Create DICOMUtils-modified.py
This is the version of DICOMUtils that contains changes used to process the sleep eeg dicom data as a pre-process for synthseg.  This contains some changes to allow selecting the specific dicom plugin to be used when loading the data.  It was modified to work in Sicer 5.2.1.

Since these changes were made, the `loadSeriesByUID` method has been further modified in Slicer core, so this code needs to be diffed and merged if it is to end up in the Slicer core.

These changes were proposed at the time since it wasn't clear that this approach is the best generalization of the needed functionality, so it is being stored here for future reference when the feature is needed again.